[refactor] url 유효성 검사 로직, 별도 클래스로 추출
